Change tests to RTL
